move scoreboard to lib
