+++ pombot/lib/scoreboard.py
@@ -1,11 +1,16 @@
+import logging
 from typing import List
 
 import discord.errors
 from discord.channel import ChannelType
+from discord.ext.commands.bot import Bot
 
 from pombot.config import Pomwars, Reactions
 from pombot.lib.messages import send_embed_message
+from pombot.state import State
 from pombot.lib.team import Team
+
+_log = logging.getLogger(__name__)
 
 
 class Scoreboard:
@@ -128,3 +133,24 @@ class Scoreboard:
                 restricted_channels.append(channel)
 
         return [full_channels, restricted_channels]
+
+
+async def setup_pomwar_scoreboard(bot: Bot):
+    """Find and remember the static scoreboard for all connected guilds."""
+    channels = []
+
+    for guild in bot.guilds:
+        for channel in guild.channels:
+            if channel.name == Pomwars.JOIN_CHANNEL_NAME:
+                channels.append(channel)
+
+    State.scoreboard = Scoreboard(bot, channels)
+    full_channels, restricted_channels = await State.scoreboard.update()
+
+    for channel in full_channels:
+        _log.error("Join channel '%s' on '%s' is not empty",
+            channel.name, channel.guild.name)
+
+    for channel in restricted_channels:
+        _log.error("Join channel '%s' on '%s' is not messagable (Missing Access)",
+            channel.name, channel.guild.name)
